ai-server/requirements.txt
flask
flask-cors
deepface
tensorflow
pillow
numpy
